+++ src/Console/ClearCacheCommand.php
@@ -5,8 +5,8 @@ namespace Nuwave\Lighthouse\Console;
 use Illuminate\Console\Command;
 use Illuminate\Contracts\Cache\Factory as CacheFactory;
 use Illuminate\Contracts\Config\Repository as ConfigRepository;
-use Illuminate\Support\Facades\File;
-use InvalidArgumentException;
+use Illuminate\Filesystem\Filesystem;
+use Nuwave\Lighthouse\Exceptions\UnknownCacheVersionException;
 
 class ClearCacheCommand extends Command
 {
@@ -14,18 +14,28 @@ class ClearCacheCommand extends Command
 
     protected $description = 'Clear the GraphQL schema cache.';
 
-    public function handle(CacheFactory $cacheFactory, ConfigRepository $config): void
+    public function handle(ConfigRepository $config): void
     {
         $version = $config->get('lighthouse.cache.version', 1);
-        if ($version === 1) {
-            $cacheFactory
-                ->store($config->get('lighthouse.cache.store'))
-                ->forget($config->get('lighthouse.cache.key'));
-        } elseif ($version === 2) {
-            $path = $config->get('lighthouse.cache.path') ?? base_path('bootstrap/cache/lighthouse-schema.php');
-            File::delete($path);
-        } else {
-            throw new InvalidArgumentException('Unknown cache version.');
+        switch ($version) {
+            case 1:
+                /** @var \Illuminate\Contracts\Cache\Factory $cacheFactory */
+                $cacheFactory = app(CacheFactory::class);
+
+                $cacheFactory
+                    ->store($config->get('lighthouse.cache.store'))
+                    ->forget($config->get('lighthouse.cache.key'));
+                break;
+            case 2:
+                /** @var \Illuminate\Filesystem\Filesystem $filesystem */
+                $filesystem = app(Filesystem::class);
+
+                $path = $config->get('lighthouse.cache.path')
+                    ?? base_path('bootstrap/cache/lighthouse-schema.php');
+                $filesystem->delete($path);
+                break;
+            default:
+                throw new UnknownCacheVersionException($version);
         }
 
         $this->info('GraphQL AST schema cache deleted.');