+++ tests/Integration/SchemaCachingTest.php
@@ -4,7 +4,6 @@ namespace Tests\Integration;
 
 use Illuminate\Contracts\Config\Repository as ConfigRepository;
 use Nuwave\Lighthouse\Schema\AST\ASTBuilder;
-use function Safe\unlink;
 use Tests\TestCase;
 use Tests\TestsSerialization;
 use Tests\Utils\Models\Comment;
@@ -13,11 +12,6 @@ class SchemaCachingTest extends TestCase
 {
     use TestsSerialization;
 
-    /**
-     * @var string
-     */
-    protected $cachePath;
-
     protected function getEnvironmentSetUp($app): void
     {
         parent::getEnvironmentSetUp($app);
@@ -25,18 +19,8 @@ class SchemaCachingTest extends TestCase
         /** @var \Illuminate\Contracts\Config\Repository $config */
         $config = $app->make(ConfigRepository::class);
         $config->set('lighthouse.cache.enable', true);
-        $this->cachePath = __DIR__.'/../storage/'.__METHOD__.'.php';
-        $config->set('lighthouse.cache.path', $this->cachePath);
-        $config->set('lighthouse.cache.version', 2);
 
         $this->useSerializingArrayStore($app);
-    }
-
-    protected function tearDown(): void
-    {
-        unlink($this->cachePath);
-
-        parent::tearDown();
     }
 
     public function testSchemaCachingWithUnionType(): void
@@ -82,8 +66,9 @@ class SchemaCachingTest extends TestCase
     protected function cacheSchema(): void
     {
         /** @var \Nuwave\Lighthouse\Schema\AST\ASTBuilder $astBuilder */
-        $astBuilder = app(ASTBuilder::class);
+        $astBuilder = $this->app->make(ASTBuilder::class);
         $astBuilder->documentAST();
+
         $this->app->forgetInstance(ASTBuilder::class);
     }
 }